Struct massa_async_pool::message::AsyncMessage

pub struct AsyncMessage {

    pub emission_slot: Slot,
    pub emission_index: u64,
    pub sender: Address,
    pub destination: Address,
    pub handler: String,
    pub max_gas: u64,
    pub fee: Amount,
    pub coins: Amount,
    pub validity_start: Slot,
    pub validity_end: Slot,
    pub data: Vec<u8>,
    pub trigger: Option<AsyncMessageTrigger>,
    pub can_be_executed: bool,
    pub hash: Hash,

}

Structure defining an asynchronous smart contract message

Fields
emission_slot: Slot

Slot at which the message was emitted

emission_index: u64

Index of the emitted message within the emission_slot. This is used for disambiguate the emission of multiple messages at the same slot.

sender: Address

The address that sent the message

destination: Address

The address towards which the message is being sent

handler: String

the handler function name within the destination address’ bytecode

max_gas: u64

Maximum gas to use when processing the message

fee: Amount

Fee paid by the sender when the message is processed.

coins: Amount

Coins sent from the sender to the target address of the message. Those coins are spent by the sender address when the message is sent, and credited to the destination address when receiving the message. In case of failure or discard, those coins are reimbursed to the sender.

validity_start: Slot

Slot at which the message starts being valid (bound included in the validity range)

validity_end: Slot

Slot at which the message stops being valid (bound not included in the validity range)

data: Vec<u8>

Raw payload data of the message

trigger: Option<AsyncMessageTrigger>

Trigger that define whenever a message can be executed

can_be_executed: bool

Boolean that determine if the message can be executed. For messages without filter this boolean is always true. For messages with filter, this boolean is true if the filter has been matched between validity_start and current slot.

hash: Hash

Hash of the message

Implementations

impl AsyncMessage

pub fn new_with_hash( emission_slot: Slot, emission_index: u64, sender: Address, destination: Address, handler: String, max_gas: u64, fee: Amount, coins: Amount, validity_start: Slot, validity_end: Slot, data: Vec<u8>, trigger: Option<AsyncMessageTrigger>, can_be_executed: Option<bool> ) -> Self

Take an AsyncMessage and return it with its hash computed

pub fn compute_id(&self) -> AsyncMessageId

Compute the ID of the message for use when choosing which operations to keep in priority (highest score) on pool overflow.

pub fn compute_hash(&mut self, for_db: bool)

Recompute the hash of the message. Must be used each time we modify one field