massa_async_pool

Module pool

Copyright (c) 2022 MASSA LABS info@massa.net This file defines a finite size final pool of asynchronous messages for use in the context of autonomous smart contracts

Structs

AsyncPool

Represents a pool of sorted messages in a deterministic way. The final asynchronous pool is attached to the output of the latest final slot within the context of massa-final-state. Nodes must bootstrap the final message pool when they join the network.

AsyncPoolDeserializer

Deserializer for AsyncPool

AsyncPoolSerializer

Serializer for AsyncPool

Constants

CAN_BE_EXECUTED_IDENT 🔒

COINS_IDENT 🔒

DESTINATION_IDENT 🔒

EMISSION_INDEX_IDENT 🔒

EMISSION_SLOT_IDENT 🔒

FEE_IDENT 🔒

FUNCTION_IDENT 🔒

FUNCTION_PARAMS_IDENT 🔒

MAX_GAS_IDENT 🔒

SENDER_IDENT 🔒

TRIGGER_IDENT 🔒

VALIDITY_END_IDENT 🔒

VALIDITY_START_IDENT 🔒