Struct massa_async_pool::AsyncMessageUpdate

pub struct AsyncMessageUpdate {

    pub emission_slot: SetOrKeep<Slot>,
    pub emission_index: SetOrKeep<u64>,
    pub sender: SetOrKeep<Address>,
    pub destination: SetOrKeep<Address>,
    pub handler: SetOrKeep<String>,
    pub max_gas: SetOrKeep<u64>,
    pub fee: SetOrKeep<Amount>,
    pub coins: SetOrKeep<Amount>,
    pub validity_start: SetOrKeep<Slot>,
    pub validity_end: SetOrKeep<Slot>,
    pub data: SetOrKeep<Vec<u8>>,
    pub trigger: SetOrKeep<Option<AsyncMessageTrigger>>,
    pub can_be_executed: SetOrKeep<bool>,

}

represents an update to one or more fields of a AsyncMessage

Fields
emission_slot: SetOrKeep<Slot>

Slot at which the message was emitted

emission_index: SetOrKeep<u64>

Index of the emitted message within the emission_slot. This is used for disambiguate the emission of multiple messages at the same slot.

sender: SetOrKeep<Address>

The address that sent the message

destination: SetOrKeep<Address>

The address towards which the message is being sent

handler: SetOrKeep<String>

the handler function name within the destination address’ bytecode

max_gas: SetOrKeep<u64>

Maximum gas to use when processing the message

fee: SetOrKeep<Amount>

Fee paid by the sender when the message is processed.

coins: SetOrKeep<Amount>

Coins sent from the sender to the target address of the message. Those coins are spent by the sender address when the message is sent, and credited to the destination address when receiving the message. In case of failure or discard, those coins are reimbursed to the sender.

validity_start: SetOrKeep<Slot>

Slot at which the message starts being valid (bound included in the validity range)

validity_end: SetOrKeep<Slot>

Slot at which the message stops being valid (bound not included in the validity range)

data: SetOrKeep<Vec<u8>>

Raw payload data of the message

trigger: SetOrKeep<Option<AsyncMessageTrigger>>

Trigger that define whenever a message can be executed

can_be_executed: SetOrKeep<bool>

Boolean that determine if the message can be executed. For messages without filter this boolean is always true. For messages with filter, this boolean is true if the filter has been matched between validity_start and current slot.