Struct massa_async_pool::AsyncPool

pub struct AsyncPool { /* private fields */ }

Represents a pool of sorted messages in a deterministic way. The final asynchronous pool is attached to the output of the latest final slot within the context of massa-final-state. Nodes must bootstrap the final message pool when they join the network.

Implementations

impl AsyncPool

pub fn new(config: AsyncPoolConfig) -> AsyncPool

Creates an empty AsyncPool

pub fn apply_changes_unchecked(&mut self, changes: &AsyncPoolChanges)

Applies pre-compiled AsyncPoolChanges to the pool without checking for overflows. This function is used when applying pre-compiled AsyncPoolChanges to an AsyncPool.

arguments

changes: AsyncPoolChanges listing all asynchronous pool changes (message insertions/deletions)

pub fn settle_slot(
&mut self,
slot: &Slot,
new_messages: &mut Vec<(AsyncMessageId, AsyncMessage)>
) -> Vec<(AsyncMessageId, AsyncMessage)>

Settles a slot, adding new messages to the pool and returning expired and excess ones. This method is called at the end of a slot execution to apply the list of emitted messages, and get the list of pruned messages for coins reimbursement.

arguments

slot: used to filter out expired messages, not stored

new_messages: list of AsyncMessage to add to the pool

returns

The list of (message_id, message) that were eliminated from the pool after the changes were applied, sorted in the following order:

expired messages from the pool, in priority order (from highest to lowest priority)

expired messages from new_messages (in the order they appear in new_messages)

excess messages after inserting all remaining new_messages, in priority order (from highest to lowest priority)

pub fn take_batch_to_execute(
&mut self,
slot: Slot,
available_gas: u64
) -> Vec<(AsyncMessageId, AsyncMessage)>

Takes the best possible batch of messages to execute, with gas limits and slot validity filtering. The returned messages are removed from the pool. This method is used at the beginning of a slot execution to list asynchronous messages to execute.

arguments

slot: select only messages that are valid within this slot

available_gas: maximum amount of available gas

returns

A vector of messages, sorted from the most priority to the least priority

pub fn get_pool_part(
&self,
last_id: Option<AsyncMessageId>
) -> Result<(Vec<u8>, Option<AsyncMessageId>), ModelsError>

Used for bootstrap Take a part of the async pool starting from the next element after last_id and with a max length of the constant ASYNC_POOL_PART_SIZE_MESSAGE_BYTES. Should always follow the same behavior as the get_ledger_part from FinalLedgermethod.

pub fn set_pool_part<'a>(
&mut self,
part: &'a [u8]
) -> Result<Option<AsyncMessageId>, ModelsError>

Set a part of the async pool. We deserialize in this function because we insert in the async pool while deserializing. Used for bootstrap.

Arguments

data: must be the serialized version provided by get_pool_part

Returns

The last id of the inserted entry (this is an optimization to easily keep a reference to the last id)