Module massa_async_pool::message

Copyright (c) 2022 MASSA LABS info@massa.net This file defines the structure representing an asynchronous message

Structs

AsyncMessage

Structure defining an asynchronous smart contract message

AsyncMessageDeserializer

AsyncMessageIdDeserializer

AsyncMessageIdSerializer

AsyncMessageInfo

AsyncMessageSerializer

AsyncMessageTrigger

Structure defining a trigger for an asynchronous message

AsyncMessageTriggerDeserializer

Deserializer for a trigger for an asynchronous message

AsyncMessageTriggerSerializer

Serializer for a trigger for an asynchronous message

AsyncMessageUpdate

represents an update to one or more fields of a AsyncMessage

AsyncMessageUpdateDeserializer

Deserializer for AsyncMessageUpdate

AsyncMessageUpdateSerializer

Serializer for AsyncMessageUpdate

Type Aliases

AsyncMessageId

Unique identifier of a message. Also has the property of ordering by priority (highest first) following the triplet: (rev(Ratio(msg.fee, max(msg.max_gas,1))), emission_slot, emission_index)